refactor: move sections in a src/sections directory
