+++ packages/web-reporter-ui/ReporterView.tsx
@@ -4,15 +4,15 @@ import {
   Measure,
   TestCaseResult,
 } from "@perf-profiler/types";
-import { CPUReport } from "./CPUReport";
-import { ReportSummary } from "./src/components/ReportSummary/ReportSummary";
-import { RAMReport } from "./RAMReport";
+import { CPUReport } from "./src/sections/CPUReport";
+import { ReportSummary } from "./src/sections/ReportSummary";
+import { RAMReport } from "./src/sections/RAMReport";
 import { AccordionSectionTitle } from "./src/components/AccordionSectionTitle";
 import Accordion from "@mui/material/Accordion";
 import AccordionDetails from "@mui/material/AccordionDetails";
 import { averageTestCaseResult } from "@perf-profiler/reporter";
 import styled from "@emotion/styled";
-import { FPSReport } from "./FPSReport";
+import { FPSReport } from "./src/sections/FPSReport";
 import { createTheme, ThemeProvider } from "@mui/material";
 import Header from "./src/components/Header";
 
@@ -21,7 +21,7 @@ import {
   IterationSelector,
   useIterationSelector,
 } from "./src/components/IterationSelector";
-import { VideoSection } from "./VideoSection";
+import { VideoSection } from "./src/sections/VideoSection";
 import { VideoEnabledContext } from "./videoCurrentTimeContext";
 
 const Padding = styled.div`
